CloudRun_jobs/requirements.txt
pandas
mlxtend
google-cloud-bigquery
loguru
db-dtypes
pandas-gbq
google-cloud-bigquery-storage
